Help me

+++ src/ollama/options.rs
@@ -4,17 +4,17 @@ use serde::Serialize;
 
 #[derive(Serialize)]
 pub struct ModelFile {
-    mirostat: Option<i32>,
+    mirostat: Option<i64>,
     mirostat_eta: Option<f32>,
     mirostat_tau: Option<f32>,
-    num_ctx: Option<i32>,
-    repeat_last_n: Option<i32>,
+    num_ctx: Option<i64>,
+    repeat_last_n: Option<i64>,
     repeat_penalty: Option<f32>,
     temperature: Option<f32>,
-    seed: Option<i32>,
+    seed: Option<i64>,
     stop: Option<String>,
-    num_predict: Option<i32>,
-    top_k: Option<i32>,
+    num_predict: Option<i64>,
+    top_k: Option<i64>,
     top_p: Option<f32>,
     min_p: Option<f32>,
 }
